added main component handler
